Sara/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
sentencepiece==0.1.99
shellingham==1.5.0
datasets==2.19.0
python-dotenv==1.0.1
numpy==1.24.4
scipy>=1.11.2
pydantic==1.10.13
openai==1.13.3
aiofiles==23.2.1
aiohttp==3.9.3
TTS==0.22
scikit-learn==1.3.2

